refactor: execution 엔티티 구조 정리
