+++ src/entities/execution/api/execute-workflow.api.ts
@@ -1,4 +1,4 @@
-import { request } from "../core";
+import { request } from "@/shared/api/core";
 
 export const executeWorkflowAPI = (workflowId: string): Promise<string> =>
   request<string>({
