+++ src/entities/execution/model/index.ts
@@ -1,4 +1,5 @@
 export * from "./execution-utils";
+export * from "./query-keys";
 export * from "./useExecuteWorkflowMutation";
 export * from "./useRollbackExecutionMutation";
 export * from "./useWorkflowExecutionQuery";
