+++ src/entities/index.ts
@@ -1,5 +1,6 @@
 export * from "./auth";
 export * from "./connection";
+export * from "./execution";
 export * from "./node";
 export * from "./oauth-token";
 export * from "./template";
