+++ src/shared/api/index.ts
@@ -1,4 +1,3 @@
 export * from "./client";
-export * from "./execution";
 export * from "./workflow";
 export * from "./query-policy";
