+++ src/shared/model/index.ts
@@ -1,3 +1,2 @@
-export * from "./execution";
 export * from "./workflow";
 export * from "./workflowStore";
