+++ src/widgets/editor-toolbar/ui/EditorToolbar.tsx
@@ -17,8 +17,10 @@ import {
   normalizeExecutionStatus,
   useExecuteWorkflowMutation,
   useRollbackExecutionMutation,
-  useSaveWorkflowMutation,
   useWorkflowExecutionsQuery,
+} from "@/entities";
+import {
+  useSaveWorkflowMutation,
   useWorkflowStore,
 } from "@/shared";
